fix: avoid warnings when compiling in Release mode
I do not fully understand why we get these warnings in the first place. Since we
never _actually_ access the 1000s component of a point. I introduce a flag to
check in Debug mode if a normal was found. In Release mode my compiler complains
that this flag is a set but ununsed variable, though it is used in inside the if
expression. I do not understand this warning either. I remedied it with this the
C++ attribute `[[maybe_unused]]`. We may discuss if this is a good thing to do.